Create P12
Create a class “Student” that would contain enrollment No, name, and gender and marks as instance variables and count as static variable which stores the count of the objects; constructors and display(). Implement constructors to initialize instance variables.
Also demonstrate constructor chaining. Create objects of class “Student” and displays all values of objects.