Create p8-2
Write a program to determine and display the maximum height reached when the ball is thrown at 5 miles / hour at an angle of 60 degrees. (Hint: Make sure to convert the initial velocity into the correct units.) The maximum height reached by a ball thrown with an initial velocity v in feet/sec at an angle of θ is given by the formula height = (0.5 * v2* sin2θ ) / 32.2. Run the program for v = 7 miles / hour and angle = 45 degrees.